Corret library bases CRF initialization; SGD Neff computation; Pseudocounts Admixture depending on Neff; Pseudocounts parallelization; Prio r on pseudocounts weights;

+++ scripts/fetch_seq.sh
@@ -11,10 +11,10 @@ DB_FROM=/data/fs03/scratch_ssd_raid/hhblits_dbs/uniprot20_16Nov10/second_round/a
 
 
 PREV=1
-for F in $DB/*; do
+for F in $DB/*a3m; do
     F=`basename $F`
     F=${F%_?\.*}
-    if [ $F != $PREV ]; then
+    if [ $F != $PREV -a ! -e $DB/$F.seq ]; then
         cp $DB_FROM/${F:0:2}/$F.a3m $DB/$F.seq
         PREV=$F
     fi
